Add SortPresenter & SortOptionPresenter
This is a refactor, to encapsulate the sort logic in separate
classes. Users should not see a change as a result of this.

The motivation for the refactor is to make it easier to change
the sort option presenter later on, and perhaps compose the
FinderPresenter to use a different sort presenter. This is
required by the stats sort logic, which is fairly specific to
one finder.

In a subsequent commit I will refresh the sort options in the
browser using only the data provided from the backend, and remove
and JS related to this logic. Then I will introduce a new class
StatisticsSortPresenter which will be used only for the stats
finder.

Trello: https://trello.com/c/IjADqrbb/591