update tests to remove conditions around small vs large number of options
This appears to be out of date code, as the current finder has ALL facets closed by default, and only open if something is selected. With the additional requirement of https://trello.com/c/5ZwqcRWh, we want to open the first facet if no facets are selected, regardless of the number of options that first facet has. I've updated the integration tests to reflect this.