Update Sort Options via Ajax calls
This moves some sort option logic out of the JS and into the
SortPresenter. It also re-renders the sort options every time
the state changes and a request is made.

The user won't notice a difference from this change.

This has a few benefits:

- we have sort logic in one place
- we can update the sort options more easily
- we can swap in different sort options dependent on the sort presenter
- when we want to switch between supergroups, we can provide different
  sorters to the frontend without a page reload

In a following commit, I will use this to change the sort options
on the fly for the stats finder using the StatisticsSortPresenter.

Trello: https://trello.com/c/IjADqrbb/591

As an aside, we could do something similar for updating facets on
the fly, too.