Use separate sort presenter for stats finder
The statistics finder has some special behaviour when sorting
I've tried to split this off from the rest of the codebase, by
introducing StatisticsSortPresenter, and providing that to the
FinderPresenter.

I also tried to avoid introducing any new JS for this logic; we
need this to work for users with JS disabled, and its better to
have the logic in one place.

The behaviour:
- we should not show public timestamp options in the sort options
  when stats (upcoming) is selected
- we should not show release timestamp options in the sort options
  when stats published or research is selected
- this behaviour should only occur on the stats finder

Trello: https://trello.com/c/IjADqrbb/591