open first facet in business finder by default
Accomplishes https://trello.com/c/5ZwqcRWh

Regardless of whether or not the first facet is selected, and regardless of whether or not any other facets are selected, we want the first facet to always remain open.

This commit makes the 'facet closed?' check explicit, by checking for the position of the facet and checking its unselected status. This information cannot live inside the model itself, so there is no benefit to trying to encapsulate it in a 'close_facet?' method.