Refactor to use MetadataPresenter for search results
Trello: https://trello.com/c/TV5xhCBX/649-give-stats-announcement-doc-types-publictimestamp-so-date-sort-works-in-faceted-search
Splitting out the metadata presentation logic from the
SearchResultPresenter. This is because we want to
compose the metadata presentation logic for each
search result independently. In a following commit we
are going to add a Stats Metadata Presenter which
will have different logic for presenting metadata.