Adds 'default_frequency' constructor param to EmailAlertSignupAPI
If omitted, the signup_url is unchanged. But if present, a 'default_frequency' URL parameter is appended to the signup_url, which is obeyed by the email-alert-frontend as of https://github.com/alphagov/email-alert-frontend/pull/461.

RSpec tests check that the parameter is correctly appended (using '?' if no other URL parameters exist, otherwise using '&').