Pass 'default_frequency' to EmailAlertSignupAPI from controller
Only want to apply a default_frequency to the Business Finder, so we explicitly check if we are the Business Finder email signup screen in signup_presenter.rb before deciding what value to return. Deliberately return 'nil' if we're not the business finder email signup, as we don't want to change the signup URL of other finders in case there are any side-effects.