Remove Mustache from advanced finder
This commit removes the two Mustache templates that are powering
part of the advanced finder.

The results page shows a count of the results. This functionality has
been shifted to the advanced finder `show` template. A
`js-result-count` hook class has been added to allow existing JS
functionality to be used when updating the count client-side.

The advanced finder results now uses the `finder/search_results`
partial. Some changes were required to tweak the advanced finder
metadata so it could be used in the finder template; it looks as if this
hasn't broken any of the other finders.