Adds StatsMetadataPresenter
The new StatsMetadataPresenter presents metadata for
stats announcements and official stats documents. By composing the
presentation logic we can present different types of metadata
for different types of documents more easily. In this specific case,
we can display a release date for announcements, and an updated date
for official stats.