Make AND condition between all facets
Attempt to build multiple queries for each 'and' facet type.
This highlights the problem of intersection with facet_values
in that facet values can't be 'anded' easily. The values from
different facets have to be queried separately and then the
intersection is calculated programmatically. This isn't performant
and should be possible in a single query.