Explicitly require addressable API
Whilst this feature worked locally and on CI in https://github.com/alphagov/finder-frontend/pull/1104, the deployment Jenkins complains of a 'NameError uninitialized constant EmailAlertSignupAPI::Addressable' which can only be resolved through requiring the library explicitly.

Work needs to be done to establish why this wasn't an issue locally or on CI, but in the meantime this hotfix should allow us to deploy.