Adds StandardJS linter to Rake task
StandardJS lints the JavaScript in Jenkins, but not when running
`bundle exec rake` locally. This PR adds `yarn run lint` to the lint
rake task so that it's easier to run all checks prior to pushing.

StandardJS is installed using Yarn, as Yarn can be used with Rails 5.1+.
Currently it's just used to make sure that StandardJS is installed with
the addition of `yarn install` to the `startup.sh` script.