Adds Chrome Headless & Karma to replace Jasmine-Rails
Jasmine-Rails relies on PhantomJS, which is abandoned. If we want to use modern JS features
such as ES6, we need to replace our test runner. The industry standard replacement for
PhantomJS is Chrome Headless.

This commit adds Chrome Headless as an environment for our test code to run within. It adds
Karma as a driver between our test code and the headless browser, and it adds Jasmine and Chai
to define assertion/helper methods.

Finally, it adds Browserify so that we are able to `require` dependencies for our specfile tests,
and karma-jquery so that we can have a global $ without having to pull it in. This meant we could
delete the copy of jquery-1.12.4 in our spec/javascripts/helpers folder.