Fix call to GOVUK.Modules ... .start()
One test consistently failed with "Unable to call method 'start' of undefined".
This was because we were calling GOVUK.modules, but it should be GOVUK.Modules (capitalised).

Changing this gave a new error message: "Object 'Modules' has no method 'start'".
Looking at the other code examples, we have to actually choose which module to call 'start' on.

Given that this function is about tracking choices, I'm fairly sure that
GOVUK.Modules.TrackQAChoices() was the intended callee.