Restore legacy values to JSON response
For backwards compatibility, we are adding these fields
back into the JSON response. These fields are no longer
in use, but this will prevent search breaking for those
users that are using live search while we deploy (and
still expect the same JSON response for the old templates).

Once we deploy, the site will continue to work for all users
that are using the site at the time of deploy. Afterwards,
we can deploy a new version with these legacy fields removed.