Official Documents Finder
This commit allows finder frontend to render a new official
documents finder. Official documents are any publication
with an attachment that is a command paper or an act paper.
The JSON has not yet been defined in Search API and
published to the content store. So an official_doc JSON fixture
is added for development. It will be removed in a later commit
once search-api/config/finders/official_documents_finder.yml
has been published.

Trello:
https://trello.com/c/6xFJqfAw/724-spin-up-finder-for-content-that-has-official-doc-status-s