Adds EmailAlertListTitleBuilder and its unit tests
We will be allowing users to sign up to mailing lists picked from
multiple facets (differing from, say, /cma-cases, which lets you
choose lists from one facet).

By default, such mailing lists would have a dynamically rendered
title that is difficult to parse:

> Find EU Exit guidance for your business with Sector / Business
> Area of Banking, markets and infrastructure, Electronics,
> Imports, and Retail, Business activity of I sell products or
> goods in the UK, I buy products or goods from abroad, and I
> sell products or goods abroad, Employ EU citizens of No, and
> Intellectual property of Trade marks

In the spirit of optimising code for deletion, we've encapsulated
this in a separate EmailAlertListTitleBuilder class, to avoid adding
complex logic to the already difficult to read EmailAlertTitleBuilder
class. In the same spirit, we've kept the Business Finder specific
overrides in the class, rather than attempt to abstract it into say
the email signup JSON schema.

Tests a number of filter scenarios and overrides for the new
'list style' email alert title.

- single-item lists should be prefixed 'following category'
- multi-item lists should be prefixed 'following categories'
- Business Finder specific overrides should be applied regardless
of how many and which other filters are being applied

Overrides should be applied to avoid the default behaviour which
can be buggy in some cases, e.g. "EU citizens of No", which we've
overwritten to "Employing non-EU citizens".

Co-authored-by: Oscar Wyatt <oscar.wyatt@digital.cabinet-office.gov.uk>