Temporarily remove email_filter_facets from fixture data
Currently this test is failing:

```
  @javascript
  Scenario: Subscribing to email alerts for business readiness finder
    When I view the business readiness finder
    And I create an email subscription
    Then I see the email subscription page
    And I cannot select any filters
```

This is functionality that we intend to remove shortly, when
we re-enable the multi-facet choice at the email signup page.

We're not ready to make the front-end change yet - so this feature
is still valid - but we'll only have to add our fixture data changes
back in when we _do_ make the front-end change.

We can then change the test to assert that the facets _are_ there.
(And then remove these changes).