Track which dynamic facets are being used
We're currently faceting on things that we don't use in the frontend.

For example, searches from https://www.gov.uk/search/all facets on `organisation`,
`manual`, `topic`, `person` and `world_location`.

This greatly increases the payload retrieved from search-api and performs
unnecessary queries in elasticsearch.

I want to do something like this:

```
def facet_query
  dynamic_facet_params = facet_params.reject { |k, _v| Services.registries.all.has_key?(k) }

  dynamic_facet_params.reduce({}) { |query, (k, v)| query.merge("facet_#{k}" => v) }
end
```

but just want to check that we know exactly which facets are being used so I
don't inadvertently break something.  Faceting isn't amazingly well covered
in the test suite, which is something we'll need to fix.

This will also make it easier for us to reintroduce dynamic faceting at a later
date as we'll be able to see it spring into life.