Don't facet on things that will be ignored
Faceting on things that are populated from local registries just adds load
to the system for no benefit.

On a search from https://www.gov.uk/search/all we go from a query like:

https://www.gov.uk/api/search.json?count=20&start=0&fields=title%2Clink%2Cdescription%2Cpublic_timestamp%2Cpopularity%2Ccontent_purpose_supergroup%2Cformat%2Cpart_of_taxonomy_tree%2Cmanual%2Corganisations%2Cpeople%2Cworld_locations&q=government&reject_link[]=%2Fsearch%2Fall&facet_manual=1500%2Corder%3Avalue.title&facet_organisations=1500%2Corder%3Avalue.title&facet_people=1500%2Corder%3Avalue.title&facet_world_locations=1500%2Corder%3Avalue.title&filter_all_part_of_taxonomy_tree[]&filter_all_part_of_taxonomy_tree[]

to

https://www.gov.uk/api/search.json?count=20&start=0&fields=title%2Clink%2Cdescription%2Cpublic_timestamp%2Cpopularity%2Ccontent_purpose_supergroup%2Cformat%2Cpart_of_taxonomy_tree%2Cmanual%2Corganisations%2Cpeople%2Cworld_locations&q=government&reject_link[]=%2Fsearch%2Fall&filter_all_part_of_taxonomy_tree[]&filter_all_part_of_taxonomy_tree[]

which is a reduction in payload from 559kb to 29.5kb, and in response time
from ~650ms to ~250ms (cold) and from ~200ms to ~25ms (warm - ie submitting
the same query).

I've added some registry stubs to feature tests because we were somehow not using
these before.

I'm all for dynamic faceting, but we should only use it when it's useful.