Use an A/B/C test
We've got some evidence that assigning people to variants isn't
foolproof, and that some people reach the app without being assigned a
session (so they automatically fall into the A bucket).  So get around
it by having:

Variant A (0% of traffic) -> A cluster
Variant B (x% of traffic) -> A cluster
Variant C (rest of traffic) -> B cluster

Some people will fall into the A bucket, but the B and C buckets
should each have about 50% of the remaining traffic.