Replace gem sass-rails with sassc-rails
This gem has been deprecated and the recommended
replacement is the sassc-rails gem.

Its a drop in replacement so we dont need to do
anything else.

For context see https://github.com/rails/sass-rails/issues/420.

It seems v6 of the gem we are using will make it
a wrapper for the other gem. We can upgrade already
so theres no need to wait for this.

Trello: https://trello.com/c/hcmzrcot/772