Remove combine_mode from business finder email signup
The combine_mode value was added to enable the use of the
OrJoinedFacetSubscriberList model when signing up to business finder
email alerts. This model has been removed from email-alert-api, so
the toggle parameter is no longer needed, since we can identify this
type of finder using the email_filter_by value.