Remove AdvancedSearch
This removes the code required to render AdvancedSearch.

AdvancedSearch is/was a generic finder for rendering
search results filtered by topic and group.
For example, search/advanced?group=guidance_and_regulation&topic=%2Feducation.
Group was a supergroup as defined in the GovukDocumentTypes
and topic was a taxon content item.

These were deprecated because we can provide the same
content via the new supergroup finders.

This was done because it'll make the codebase easier
to work with, and provide a more consistent experience
for end users when they use finders.

We are taking extra steps to ensure that the deprecation
happens gracefully. Advanced search links should continue
to work: they will redirect to the appropriate supergroup
finder. The blue header at the top of topic pages will
be added to the top finders when users come from a topic
page, and the topic facet will be hidden.