Add custom options question and change eu employment question
The `custom_options` are a way to add radio list with custom labels.
Previously, the labels were taken from the facet values which did
not always make sense based on the questions in the QA view.

It is important to note that each value must match a facet value
published by content tagger which it will redirect to, otherwise, it would not be
applied and ignored by the finder. These lists are not validated against
the content item, therefore, if the content item facet values change, this will need
to be changed too.

Trello:
https://trello.com/c/WDUH4kll/186-change-wording-for-eu-citizens