Add Enhanced Ecommerce tracking
This adds tracking attributes to elements on search
pages. These new attributes will be used by
GOVUK.Ecommerce that is provided by static:
https://github.com/alphagov/static/blob/master/app/assets/javascripts/analytics/ecommerce.js

data-analytics-ecommerce is added to the container
data-ecommerce-start-index=start_offset is also added to the container
data-list-title is too, this is the finder name
data-search-query is too, this is the search query

results now have an attribute data-ecommerce-row,
which signals this is a row we should track.

Results have a path and content id, which means
we now fetch a content_id as part of the return fields
from Search API.

With these attributes, the Ecommerce component
will set up a bunch of event handlers, and then
send tracking information to Google Analytics.

We call GOVUK.Ecommerce.start() after every
new search. The intent is to make the component
set up listeners on the new elements when a
new search is created.

We now send two requests to GA when a user clicks
a link. Does anyone know if we can make just one
call to GA? From the looks of things it's not
possible, but if we can I'd like to!

Note: This should be merged after https://github.com/alphagov/static/pull/1787
is deployed.

Trello: https://trello.com/c/YXkJcbBo/851.