Add SearchAnalytics component
This will serve as a wrapper module for our use of
GOVUK.analytics, a globally defined (but sometimes
undefined) object.

We can also throw in new tracking logic (such as
eCommerce tracking) here when we add it.

This is a bugfix, as GOVUK.analytics can be undefined
and we didn't guard against that.

Trello: https://trello.com/c/Ebmxi4AN/871