Translate slugs to content ids for email alert signup
Signing up to content linked to organisations, people and
world locations should result in a SubscriberList being created with
parameters included in the 'link' field of a SubscriberList.

Each of these parameters are given as slugs and should be converted to
a content_id, except for part_of_taxonomy_tree which is already a
content_id.