Strip excess whitespace from parameters
We see a lot of requests to the business readiness finder where the
last facet value in the URL has a trailing space.  There's probably a
widely circulated bad link, or page with the bad link.  Rather than
track down the origin, we can make finder-frontend gracefully handle
user errors like this.