Research and Stats Finder displays older stats announcements
Prior to this commit, the stats finder returns stats
announcements due to be published from today onwards,
when 'Upcoming Statistics' is selected. Now it will
return those that were due to be released any time
in the previous month, provided the stats haven't yet
been published. This fixes a bug that was caused
when stats are due to be published and then cancelled,
resulting in the announcement disappearing from the finder.