Hide 'Show only Brexit results' if filtering on the Brexit Topic.
When we navigate from the Brexit Topic Page, we set the 'topic'
parameter to the Brexit content_id which filters the results
accordingly.

This means that the checkbox 'Show only Brexit results' should
be removed because selecting or deselecting it will have no
effect and is likely to confuse users.