Add a development page for finder-frontend
Currently this app doesn't have anything on the root, which sometimes
causes problems (the dreaded "no such method `.to_html` for nilClass").

This solves that by adding a simple homepage, which is only visible on
Heroku and development (because in production nothing will route to it).

The page is currently a list of stuff that's rendered by the app,
according to the search-api. This is useful to find things to test.