Show 'lead times' instead of 'due dates'
Previously we showed a specific due date for each action. This replaces
the dates with an textual lead time e.g. '3 months'.

This also fixes the feature spec to actually check which actions are
shown, as the associated method was not previously being called. In order
to mitigate the increasing number of feature steps, this combines the
question steps together, since the flow is custom for each scenario.