Add some validations for checklist YAML files
This adds a few validations to check the content of the checklist YAML
files at the point of loading: required fields are present and
referenced criteria exist. Other optional fields could still be written
incorrectly, but that doesn't seem as critical to test. We should add
further validations as new fields are added to the checklist classes.