Add priority field to sort actions by
This adds a new priority field and validates it is 'High', 'Medium' or
'Low' for each action. Further sorting is possible within priorities by
reordering actions with the same priority in the YAML file.

This converts the textual priorities to numerical ones at the poin of
sorting, which has the advantage that the interpretation of the priority
field is clear. We could move to numerical priorities if a more
fine-grained scale of priorities becomes desirable.