Add PageService to calculate pages in Checklists to avoid redirects
Calculating which page to show next was done as a helper inline in the
controller. This functionality has been extracted to a separate model.

This also allows us to calculate the next viewable page directly
without having to redirect multiple times.

The Paging service also keeps track of the next page and whether
to redirect