Tweak checklist routes to use new prefix
This changes the question and results checklist routes to use the new
'/get-ready-brexit-check/' prefix, introducing path aliases for both to
make any future changes easier. We can't use the the prefix route on its
own, since this is now reserved for the start page, which is going to be
published via Mainstream Publisher.