Add support for 'single wrapped' question types
This adds support for a single/radio type question that conditionally
revelas a set of checkboxes when a radio selection is made.

Previously we formatted options for radios/checkboxes as part of the
Question class. This moves the formatting concern to the checklist
helper and modifies it to allow formatting of individual options to
support the conditional reveal feature.

Since the possible criteria for a question may now be nested, this
updates the Question#possible_criteria method to still be accurate.