Add questions/criteria for businesses
This adds questions and matching criteria for businesses, copying
exactly those currently used by the business finder.

Since we don't currently have a question type that groups related
checkboxes, those for the sector question are presented as one long
list, with comments for the proposed groupings. We will implement
support for grouping checkboxes in future work.

Previously the feature test duplicated the code to answer each
question. This refactors the feature to add helper methods for answering
questions and checking actions exist, as well as reforming the scenarios
to focus on the two main 'business' and 'citizen' use cases.