Temporarily remove criteria dependencies
Previously we supported having dependencies between criteria to cope
with the user changing their answers and making certain questions (and
thus criteria) unavailable. This removes the dependency capability as
the existing implementation is not powerful enough to express the
dependencies for citizen questions.

The previous approach also admitted a bug when signing up for emails,
since the subscriber list would also refer to criteria that may no
longer be relevant e.g. if the user changes to a "No" response after
checking some options with the 'single wrapped' question type.