Use yes/no questions instead of none/neither
Previously we had a couple of multiple choice questions with a 'none',
'other' or 'neither' option, which is confusing because the user can
also check other options that logically conflict with this. This changes
the approach for such questions to remove such options, so the user just
doesn't make any selection. In the case of pets/driving, it also makes
sense to merge this question using the 'single wrapped' question type.