Re-use criteria logic for questions
This replaces the current 'AND' logic for question dependencies with the
same logic we use for actions. In order to support that a question
without any dependencies will be shown, this changes the default return
value for criteria logic to be truthy, while changing the validation for
actions to force them to have criteria, which seems to be the intention
of the original return value.