Stop caching checklist YAML in dev
This moves the memoisation of questions/criteria into the associated
model classes, which gives us the opportunity to auto-reload them in
development and quicken the debug/test cycle when changing them.

Previously we switched the behaviour of loading actions to potentially
include ones that are soft deleted. This temporarily removes the soft
deletion behaviour, which isn't currently used, in order to get the same
(lack of) memoisation benefits for actions in development.