Add back the criteria dependencies
Previously we removed the capability for a user to change
the parameters when completing a dynamic lists. Instead
when the user wants to update some selections they
would have to start again from scratch.

This was needed because we needed a way to disregard
previous selections. When displaying the results only
the new answers should be presented to the user and
not any possibe leftover selections that are no
longer valid.

This commit adds a way of filtering all criteria submitted
through the URL by selection only those criteria that
the user selected in their latest run

related: https://github.com/alphagov/finder-frontend/pull/1386