Add variance to the cache refresh period
At present this refreshes hourly.  If we greatly increase the number of machines
this might present an unhelpful load all at once to a number of services.

By adding some semblance of variance, we're less likely to all be calling the
same things at the same time except on deploy.