Add job to populate registry caches if empty
At present we refresh all registry caches on deploy.  This is generally
OK, but if we increase the number servers dramatically then this might
cause load issues down the line.

It's mostly unnecessary anyway.  Mostly we just want to ensure that we
have _something_ in the cache, not that it's exactly up to date.

The new job will populate the caches if they're empty, but otherwise
leave well alone.