Use one 'lib/' directory instead of two
Previously we had some code in 'app/lib/' and other code in 'lib/',
which is really confusing because it's unclear where a particular file
lives and is not accurately reflected in the single 'spec/lib' dir. This
moves all the 'app/lib/' files into 'lib/' and autoloads all of them.