Healthcheck to make sure Action URLs are not broken
This commit introduces a simple script that pings every url in the
actions yaml for a header response of 200:OK. If the URL does not
respond with 200:OK, it will report the failing URL and the
response code for manual checking.

This script should be habitually run as a production process.