Iterate CSV rake task test
Previously we (repeatedly) loaded individual tasks as part of the spec.
This switches to loading all tasks once, which supports the further work
we will do to add more rake tasks/tests. This also modifies the rake task
to import questions, etc. to stop it leaking into the test output.