Add GoogleDriveCsvDownloader
This takes a Google Sheet id, downloads the first tab of the sheet to a CSV and
writes it to a given location.

The code for this has been adapted from the
Google Drive API quickstart script:
https://developers.google.com/drive/api/v3/quickstart/ruby

Google::Auth::ClientId is expecting to read a configuration JSON file so
developers looking to download a CSV from the API will need to follow the
instructions in the quickstart guide above to get the JSON file.