Iterate change note fields
This modifies the fields we accept for change notes in preparation for
them being used in notifications. Previously we:

   - Referenced actions or questions
   - Used unstructured, textual IDs
   - Had fields like 'title' and 'text'
   - Did not categorise the change notes
   - Had freeform dates

Future change notes:

   - Are either an 'addition' or a (major) 'content_change'
   - Have the necessary fields for sending notifications
   - Have machine readable date/times to support reformatting
   - Have unique UUIDs for compatibility with Email Alert API

Since we only intend to send notifications about changes to actions,
it's no longer necessary to support references to questions.

Since none of the existing change notes are valid with the new fields,
this comments all of them out pending future work to fix them.