Move valid? method code into specs
Previously we did had validations in both the specs and the models for
questions and actions, which is confusing since e.g. 'Question#valid?'
only validates the question criteria. This moves the 'valid?' code into
the specs for questions and actions, but leaves the actual validation of
criteria cohesively together with other CriteriaLogic classes.