Add hint text and Ireland option to travel question
This is a content update to help clarify the travel question. Currently
everyone will be presented with this question and it confusing for
people living in the UK to be asked if they are traveling to the UK.
The hint text helps clarify that people may be in this situation if
they had began travelling before 31st of Oct. We also need to start
capturing people traveling to ireland seperately.