Add rake task to download and convert CSV from Google Drive
This adds a rake task that makes use of
`Checklists::ConvertCsvToYaml::GoogleDriveCsvDownloader` to download the latest
version of the Actions CSV from Google Drive and convert it to actions.yaml.
This task is intended to be run locally before committing the changes made to
actions.yaml.  Developers running this task will need to add the ID of the
Google sheet to a .env file in order to provide the Google Drive API with the
right spreadsheet to download.