Introduce factories for a psuedo-model classes
Previously we manually created instances of questions, etc. for testing,
which had a few issues:

   - There is no way to validate the test data
   - The test data has to conform to the YAML

Introducing validations into the classes right now would be a pain, as
all the test data would need to be modified to make it valid. To support
the introduction of validations, this creates factories for questions
and actions, and modifies they way all checklist psuedo-models are
instianted to separate creating an instance and loading from YAML.