Separate syntactic and semantic validations
Previously we used the specs for questions, etc. as a proxy to validate
the data in our YAML files is in the right format (syntactically valid)
and that references between YAML files work (semantically valid). In
order to make our factories more robust, this uses ActiveModel to move
syntactic validations into the classes under test, and aggregates the
remaining semantic checks into a single 'data integrity spec', since we
don't want them to run for each instance of a question, etc., and we
also don't want them to apply for questions, etc. created by factories.