Add tracking for action and guidance links
Changes action index to be formatted in the following format: `1.01` or `1.03` so that when
the index is `1.10` the ordering on Google Analytics is done correctly.

Why do we track position of links?

The current reasoning for tracking position of links is to understand where users are clicking on links.
This information will be used to make informed decisions about ordering of links and understand how far
down the list users scroll and click. This is also something that is done in other finders across GOV.UK.

This is something that will need to be reviewed in the future and understand the user need more clearly.