Pre-warm content items in cache
We'd like to prevent requests from users requiring a content store call.
This ensures that content items finder-frontend renders are cached ahead
of any users requests, and are periodically refreshed. There are also
rake tasks we can call to refresh the cache during a deploy and periodically.

This is a similar pattern to the registries.

The reason we're doing this now is because there are more finder-frontend
machines, which makes it more likely for an end user to hit a cold cache.

Trello: https://trello.com/c/1PG0oqZg/992-warm-cache-with-finder-content-items