Change change note time -> date
Previously each change note had a 'time' field, which was confusing
because, while the date is intuitive, it's unclear what the exact time
of a given change note is referring to. This renames and reformats the
'time' field to simply store the date of the change, which is consistent
with the latest copy for the emails we're intending to send.