Iterate notification emails again
This updates the Brexit checker emails to reflect the latest design
iteration. We decided to use a static title/subject for all emails to
better reflect the checklist results page as a single 'content item' on
GOV.UK, rather than trying to hack-in action-specific messaging.

Part of the new design is to have a static piece of copy after the (now
static) title, which introduces the type of change, so that the final
('Changed') section is now exclusively for showing a change note.