Enforce consistent quote style
Previously we allowed any style of quotation between files and within
the same file, which only serves to increase the mental load when trying
to write consistent code. It would be quite a large change to reenable
the associated RuboCop across the whole of GOV.UK, but this change will
help reduce the quotation choice overhead for at least this repo.