Add group_id when creating Brexit subscriber lists
This adds an id that can be used to group together all the different
lists that are created for the Get ready for Brexit checker. This can be
used as a means to query data for all the lists that are variants.

This also updates the feature test to verify that the create request has
performed with the parameters we anticipate. There were a few of these I
had to resolve.