Highlight search terms in result descriptions
This will make search terms in the search result
descriptions appear bold.

We had this behaviour before, which is why no additional
work in search-api was required, but it was disabled when
we made a change to search (not sure whether this was intended).

This will highlight terms in all finders.