Add previous_question_index helper
This helper determines what the index of a previous question of a given
question is, depending on the criteria that have been supplied. This is
intended to be used in order to determine which question to direct a user to
when they click a 'Back' button on a question page, which will be implemented
in a future commit.