Add email tracking for dynamic list emails
Currently we send out notifications without any tracking on the links they contain. A given notification may have up to 2 links: the title of the action; and a link to some guidance. Note that some links are to external websites and we won't be able to track clicks on those.

After talking to Paul Cronk, we propose to manually inject the following query params into links that go to gov.uk:

   - utm_source = <uuid of the change note>
   - utm_medium = email
   - utm_campaign = govuk-brexit-checker

Normally we also set `utm_content = <subscription frequency>`, but we don't have the data to do that upfront. We've also changed the campaign from the normal `govuk-notifications`, which we should bear in mind in for other links in the email e.g. the link to the results list.

Normally an 'Update on GOV.UK' has a title link with tracking attributes automatically injected into it e.g.

https://www.gov.uk/government/news/government-backs-cutting-edge-tech-to-drive-down-shipping-emissions?utm_source=cb68c5c0-974f-46e3-a1db-2287917b17cd&utm_medium=email&utm_campaign=govuk-notifications&utm_content=immediate

Currently our notifications don't have a (title) link, and instead have other custom links in the freeform body, which don't benefit from this automatic injection mechanism.

https://trello.com/c/5UPZU8jc/242-collect-analytics-for-when-users-click-on-action-links-in-notifications