Rename ChangeNote to Notification
Change notes have been used in the application as a log of changes we will notify subscribers about.
Each ones also has an optional 'note' that content can use to describe the nature of the change to subscribers.
This has generated historic confusion for developers and makes these concepts difficult for new starters to instantly grasp.