Change text of do-you-own-a-business question
Recent research has shown people are confused by the initial business question, and are selecting 'no' when they own a business, but may not feel like they are 'preparing it for Brexit'. We should change this question so that more businesses see it as relevant to their business (even if tiny), and so that it’s relevant even if they aren’t preparing right now.