Fix flourinated-gas registration service url
The link without the path works, but the user is redirected to /register/index, however we appear to be appending google analytics ids as optional params to URLS when making this hop. If the param is appended to the root it causes a failure of the gas registration app. However if the param is appended to /register, everything continues to work